requirements.txt
Authlib~=0.15.4
apscheduler==3.8.1
connexion[swagger-ui]~=2.11.2
dramatiq[redis,watch]==1.12.3
email-validator~=1.1.3
boto3~=1.24.80
Bcrypt-Flask==1.0.2
giturlparse~=0.10.0
click-option-group~=0.5.3
flask-cors==3.0.10
flask-marshmallow~=0.14.0
flask-restful==0.3.9
flask-login~=0.5.0
flask-shell-ipython==0.4.1
flask-wtf~=1.0.0
Flask-APScheduler==1.12.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Mail~=0.9.1
Flask~=2.0.3
gunicorn~=20.1.0
itsdangerous~=2.1.0
jwt==1.3.1
loginpass==0.5
marshmallow-sqlalchemy~=0.27.0
nf-core~=2.5.0
prometheus-flask-exporter~=0.20
pygit2~=1.9.0
psycopg2-binary~=2.9.1
pyopenssl==22.0.0
pytest-mock~=3.7.0
pytest~=7.0.1
python-dotenv~=0.19.0
python-jenkins==1.7.0
python-redis-lock~=3.7.0
PyGithub~=1.55
PyYAML~=5.4.1
pika~=1.2.0
redis~=4.1.4
requests~=2.27.1
snakemake==6.13.1
networkx~=2.8.2
rich~=12.4.1
rocrate==0.6.0
SQLAlchemy~=1.3.24
wheel~=0.37.1
Werkzeug~=2.0.0
repo2rocrate~=0.1.1
urllib3==1.26.11
